\name{SwissInhabitants}
\alias{SwissInhabitants}
\alias{SwissInhabitants1900}
\docType{data}
\title{SwissInhabitants in 1900}
\description{
Number of inhabitants in the 2896 villages of Switzerland in the year 1900.
}
\format{
A numeric vector of length 2896 containing the population counts.
}
\usage{data("SwissInhabitants")}
\details{
This data set consists of the number of inhabitants in the 2896 communes 
(cities and villages) in Switzerland in 1900. The data is unordered for 
anonymity reasons.
}
\source{
Schuler, M., Ullmann, D. (2002). Eidgenossische Volkszahlung: 
Bevoelkerungsentwicklung der Gemeinden. Bundesamt fur Statistik, 
Neuchatel, Switzerland.
}
\references{
Behnisch, M., Ultsch, A. (2010). Population Patterns in Switzerland 1850-2000. 
In: Gaul, W. et al (Eds), Advances in Data Analysis, Data Handling and 
Business Intelligence, Springer, Heidelberg, pp. 163-173.
}
\examples{
data(SwissInhabitants)
summary(SwissInhabitants)
}
\keyword{datasets}
