% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cABC_analysis.R
\name{cABC_analysis}
\alias{cABC_analysis}
\title{ABC Classification}
\usage{
cABC_analysis(Data, ABCcurvedata, PlotIt = FALSE, useGGPlot = TRUE)
}
\arguments{
\item{Data}{Positive numeric vector which is not uniformly distributed.
If matrix or dataframe then the first column will be used.}

\item{ABCcurvedata}{Legacy: only for internal usage, Optional list returned by 
ABCcurve(), containing the interpolated ABC curve.}

\item{PlotIt}{Logical. If TRUE, an ABC plot is generated.}

\item{useGGPlot}{Logical, default TRUE. Sets if base R or gg plot should be used}
}
\value{
A list containing: 
 Aind, Bind, Cind: Indices of the items partitioned into classes A, B, and C.
   If the A class is empty, the B class is set as the A class.
 ABexchanged: Logical; TRUE if the Pareto point and Break-even point were 
   swapped to maintain coordinate logic.
 A, B, C: The c(x, y) coordinates for the Pareto point (A), 
   the Break-even point (B), and the Submarginal point (C).
 smallestAData: The cumulative yield at the boundary of Class A.
 smallestBData: The cumulative yield at the boundary of Class B.
 AlimitIndInInterpolation: The index of the A boundary in the [p, ABC] curve.
 BlimitIndInInterpolation: The index of the C boundary in the [p, ABC] curve.
 p: Numeric vector of the effort (x-axis) of the interpolation curve.
 ABC: Numeric vector of the yield (y-axis) of the interpolation curve.
 ABLimit: The data point closes to the value threshold separating Class A from Class B.
 BCLimit: The data point closes to the value threshold separating Class B from Class C.
}
\description{
Divides a numeric dataset into three classes (A, B, and C) using
ABC analysis. The classification is based on geometric properties
of the ABC curve and identifies regions of high, balanced, and
low efficiency.
}
\details{
Class interpretation:
A = low effort, high yield 
B = balanced effort and yield
C = high effort, low yield
}
\author{
AH (01/2026): Refactor, Group assignment fixed, monotone Spline
}
